Pipeline/finetuning_server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.1
pydantic==2.11.4
redis==5.2.1
runpod==1.7.7
python-dotenv==1.1.0
